+++ packages/editor/editor-core/src/plugins/rule/index.tsx
@@ -4,6 +4,7 @@ import { EditorPlugin } from '../../types';
 import { messages } from '../insert-block/ui/ToolbarInsertBlock';
 import keymapPlugin from './pm-plugins/keymap';
 import inputRulePlugin from './pm-plugins/input-rule';
+
 import {
   addAnalytics,
   ACTION,
@@ -13,8 +14,14 @@ import {
   EVENT_TYPE,
 } from '../analytics';
 import { IconDivider } from '../quick-insert/assets';
+import { safeInsert } from '../../utils/insert';
+import { Transaction } from 'prosemirror-state';
+import { Fragment } from 'prosemirror-model';
+import { getEditorProps } from '../shared-context';
 
 const rulePlugin = (): EditorPlugin => ({
+  name: 'rule',
+
   nodes() {
     return [{ name: 'rule', node: rule }];
   },
@@ -44,7 +51,25 @@ const rulePlugin = (): EditorPlugin => ({
           <IconDivider label={formatMessage(messages.horizontalRule)} />
         ),
         action(insert, state) {
-          const tr = insert(state.schema.nodes.rule.createChecked());
+          let tr: Transaction<any> | null = null;
+          const { allowNewInsertionBehaviour } = getEditorProps(state);
+          if (allowNewInsertionBehaviour) {
+            /**
+             * This is a workaround to get rid of the typeahead text when using quick insert
+             * Once we insert *nothing*, we get a new transaction, so we can use the new selection
+             * without considering the extra text after the `/` command.
+             **/
+            tr = insert(Fragment.empty);
+            tr = safeInsert(
+              state.schema.nodes.rule.createChecked(),
+              tr.selection.from,
+            )(tr);
+          }
+
+          if (!tr) {
+            tr = insert(state.schema.nodes.rule.createChecked());
+          }
+
           return addAnalytics(tr, {
             action: ACTION.INSERTED,
             actionSubject: ACTION_SUBJECT.DOCUMENT,