+++ packages/editor/editor-core/src/ui/RecentSearch/withActivityProvider.tsx
@@ -4,8 +4,8 @@ import {
   ProviderFactory,
   WithProviders,
   Providers,
+  Diff,
 } from '@atlaskit/editor-common';
-import { Diff } from '../../utils/types';
 
 export interface ExpandedActivityProviderProps {
   providerFactory: ProviderFactory;