+++ packages/editor/renderer/src/ui/Renderer/index.tsx
@@ -30,7 +30,7 @@ import { ACTION, ACTION_SUBJECT, EVENT_TYPE } from '../../analytics/enums';
 import { AnalyticsEventPayload, PLATFORM, MODE } from '../../analytics/events';
 import AnalyticsContext from '../../analytics/analyticsContext';
 import { CopyTextProvider } from '../../react/nodes/copy-text-provider';
-
+import { Provider as SmartCardStorageProvider } from '../SmartCardStorage';
 export interface Extension<T> {
   extensionKey: string;
   parameters?: T;
@@ -60,6 +60,7 @@ export interface Props {
 export class Renderer extends PureComponent<Props, {}> {
   private providerFactory: ProviderFactory;
   private serializer?: ReactSerializer;
+  private rafID: number | undefined;
 
   constructor(props: Props) {
     super(props);
@@ -68,14 +69,15 @@ export class Renderer extends PureComponent<Props, {}> {
     startMeasure('Renderer Render Time');
   }
 
-  onRenderComplete() {
+  private anchorLinkAnalytics() {
     const anchorLinkAttributeHit =
       !this.props.disableHeadingIDs &&
       window.location.hash &&
-      document.getElementById(window.location.hash.slice(1));
+      document.getElementById(
+        decodeURIComponent(window.location.hash.slice(1)),
+      );
 
     if (anchorLinkAttributeHit) {
-      anchorLinkAttributeHit.scrollIntoView();
       this.fireAnalyticsEvent({
         action: ACTION.VIEWED,
         actionSubject: ACTION_SUBJECT.ANCHOR_LINK,
@@ -86,9 +88,6 @@ export class Renderer extends PureComponent<Props, {}> {
   }
 
   componentDidMount() {
-    // add setTimeout to ensure the rendering process has completed.
-    window.setTimeout(() => this.onRenderComplete());
-
     this.fireAnalyticsEvent({
       action: ACTION.STARTED,
       actionSubject: ACTION_SUBJECT.RENDERER,
@@ -96,7 +95,7 @@ export class Renderer extends PureComponent<Props, {}> {
       eventType: EVENT_TYPE.UI,
     });
 
-    requestAnimationFrame(() => {
+    this.rafID = requestAnimationFrame(() => {
       stopMeasure('Renderer Render Time', duration => {
         this.fireAnalyticsEvent({
           action: ACTION.RENDERED,
@@ -117,6 +116,7 @@ export class Renderer extends PureComponent<Props, {}> {
           eventType: EVENT_TYPE.OPERATIONAL,
         });
       });
+      this.anchorLinkAnalytics();
     });
   }
 
@@ -204,12 +204,14 @@ export class Renderer extends PureComponent<Props, {}> {
                   this.fireAnalyticsEvent(event),
               }}
             >
-              <RendererWrapper
-                appearance={appearance}
-                dynamicTextSizing={!!allowDynamicTextSizing}
-              >
-                {result}
-              </RendererWrapper>
+              <SmartCardStorageProvider>
+                <RendererWrapper
+                  appearance={appearance}
+                  dynamicTextSizing={!!allowDynamicTextSizing}
+                >
+                  {result}
+                </RendererWrapper>
+              </SmartCardStorageProvider>
             </AnalyticsContext.Provider>
           </IntlProvider>
         </CopyTextProvider>
@@ -234,6 +236,10 @@ export class Renderer extends PureComponent<Props, {}> {
 
   componentWillUnmount() {
     const { dataProviders } = this.props;
+
+    if (this.rafID) {
+      window.cancelAnimationFrame(this.rafID);
+    }
 
     // if this is the ProviderFactory which was created in constructor
     // it's safe to destroy it on Renderer unmount