+++ packages/media/media-card/src/root/card/index.tsx
@@ -1,6 +1,15 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { Component } from 'react';
+
+import {
+  AnalyticsContext,
+  UIAnalyticsEvent,
+  withAnalyticsEvents,
+  withAnalyticsContext,
+  WithAnalyticsEventsProps,
+} from '@atlaskit/analytics-next';
+import DownloadIcon from '@atlaskit/icon/glyph/download';
 import {
   MediaClient,
   FileDetails,
@@ -11,28 +20,53 @@ import {
   isDifferentIdentifier,
   isImageRepresentationReady,
   addFileAttrsToUrl,
+  FileState,
+  ExternalImageIdentifier,
 } from '@atlaskit/media-client';
-import DownloadIcon from '@atlaskit/icon/glyph/download';
-import { AnalyticsContext, UIAnalyticsEvent } from '@atlaskit/analytics-next';
+import { MediaViewer, MediaViewerDataSource } from '@atlaskit/media-viewer';
+
 import { Subscription } from 'rxjs/Subscription';
 import { IntlProvider } from 'react-intl';
-import { MediaViewer, MediaViewerDataSource } from '@atlaskit/media-viewer';
-import { CardAction, CardDimensions, CardProps, CardState } from '../..';
-import { CardView } from '../cardView';
+import {
+  CardAction,
+  CardDimensions,
+  CardProps,
+  CardState,
+  CardStatus,
+} from '../..';
+import { CardView, CardViewBase } from '../cardView';
 import { LazyContent } from '../../utils/lazyContent';
 import { getDataURIDimension } from '../../utils/getDataURIDimension';
 import { getDataURIFromFileState } from '../../utils/getDataURIFromFileState';
 import { extendMetadata } from '../../utils/metadata';
 import { isBigger } from '../../utils/dimensionComparer';
-import { getCardStatus } from './getCardStatus';
-import { InlinePlayer } from '../inlinePlayer';
+import {
+  getCardStatus,
+  getCardStatusFromFileState,
+  getAnalyticsStatusFromCardStatus,
+  getCardProgressFromFileState,
+  getAnalyticsErrorStateAttributes,
+  AnalyticsErrorStateAttributes,
+  AnalyticsLoadingAction,
+} from './getCardStatus';
+import { InlinePlayer, InlinePlayerBase } from '../inlinePlayer';
 import {
   getUIAnalyticsContext,
   getBaseAnalyticsContext,
+  createAndFireCustomMediaEvent,
+  getFileAttributes,
 } from '../../utils/analytics';
 
-export class Card extends Component<CardProps, CardState> {
+export type CardWithAnalyticsEventsProps = CardProps & WithAnalyticsEventsProps;
+export class CardBase extends Component<
+  CardWithAnalyticsEventsProps,
+  CardState
+> {
   private hasBeenMounted: boolean = false;
+  private lastAction?: AnalyticsLoadingAction = undefined;
+  private lastErrorState?: AnalyticsErrorStateAttributes = {};
+  private resolvedId: string = '';
+  cardRef: React.RefObject<CardViewBase | InlinePlayerBase> = React.createRef();
 
   subscription?: Subscription;
   static defaultProps: Partial<CardProps> = {
@@ -49,10 +83,47 @@ export class Card extends Component<CardProps, CardState> {
     isPlayingFile: false,
   };
 
+  // we add a listener for each of the cards on the page
+  // and then check if the triggered listener is from the card
+  // that contains a div in current window.getSelection()
+  // won't work in IE11
+  onCopyListener = () => {
+    if (typeof window.getSelection === 'function') {
+      const selection = window.getSelection();
+
+      if (
+        this.cardRef.current &&
+        this.cardRef.current.divRef.current instanceof Node &&
+        selection &&
+        selection.containsNode(this.cardRef.current.divRef.current, true)
+      ) {
+        this.fireAnalytics();
+      }
+    }
+  };
+
+  fireAnalytics = async () => {
+    const { createAnalyticsEvent, identifier } = this.props;
+
+    createAndFireCustomMediaEvent(
+      {
+        eventType: 'ui',
+        action: 'copied',
+        actionSubject: 'file',
+        actionSubjectId:
+          identifier.mediaItemType === 'file'
+            ? await identifier.id
+            : 'external-image',
+      },
+      createAnalyticsEvent,
+    );
+  };
+
   componentDidMount() {
     const { identifier, mediaClient } = this.props;
     this.hasBeenMounted = true;
     this.subscribe(identifier, mediaClient);
+    document.addEventListener('copy', this.onCopyListener);
   }
 
   UNSAFE_componentWillReceiveProps(nextProps: CardProps) {
@@ -91,6 +162,7 @@ export class Card extends Component<CardProps, CardState> {
     this.hasBeenMounted = false;
     this.unsubscribe();
     this.releaseDataURI();
+    document.removeEventListener('copy', this.onCopyListener);
   }
 
   releaseDataURI = () => {
@@ -114,33 +186,67 @@ export class Card extends Component<CardProps, CardState> {
     }
   };
 
-  async subscribe(identifier: Identifier, mediaClient: MediaClient) {
+  subscribe(identifier: Identifier, mediaClient: MediaClient) {
     const { isCardVisible } = this.state;
+
     if (!isCardVisible) {
       return;
     }
 
     if (identifier.mediaItemType === 'external-image') {
-      const { dataURI, name } = identifier;
-
-      this.setState({
-        status: 'complete',
-        dataURI,
-        metadata: {
-          id: dataURI,
-          name: name || dataURI,
-          mediaType: 'image',
-        },
-      });
-
-      return;
+      this.subscribeExternalFile(identifier);
+    } else {
+      this.subscribeInternalFile(identifier, mediaClient);
     }
+  }
 
+  subscribeExternalFile(identifier: ExternalImageIdentifier) {
+    const { createAnalyticsEvent } = this.props;
+    const { dataURI, name } = identifier;
+    this.resolvedId = dataURI;
+
+    createAndFireCustomMediaEvent(
+      {
+        eventType: 'operational',
+        action: 'commenced',
+        actionSubject: 'mediaCardRender',
+        actionSubjectId: dataURI,
+      },
+      createAnalyticsEvent,
+    );
+
+    this.setState({
+      status: 'complete',
+      dataURI,
+      metadata: {
+        id: dataURI,
+        name: name || dataURI,
+        mediaType: 'image',
+      },
+    });
+  }
+
+  async subscribeInternalFile(
+    identifier: FileIdentifier,
+    mediaClient: MediaClient,
+  ) {
     const { id, collectionName, occurrenceKey } = identifier;
-    const resolvedId = typeof id === 'string' ? id : await id;
+    const { createAnalyticsEvent } = this.props;
+    this.resolvedId = await id;
+
+    createAndFireCustomMediaEvent(
+      {
+        eventType: 'operational',
+        action: 'commenced',
+        actionSubject: 'mediaCardRender',
+        actionSubjectId: this.resolvedId,
+      },
+      createAnalyticsEvent,
+    );
+
     this.unsubscribe();
     this.subscription = mediaClient.file
-      .getFileState(resolvedId, { collectionName, occurrenceKey })
+      .getFileState(this.resolvedId, { collectionName, occurrenceKey })
       .subscribe({
         next: async fileState => {
           let {
@@ -160,7 +266,7 @@ export class Card extends Component<CardProps, CardState> {
             dataURI = src;
             if (dataURI && contextId) {
               dataURI = addFileAttrsToUrl(dataURI, {
-                id: resolvedId,
+                id: this.resolvedId,
                 collection: collectionName,
                 contextId,
                 mimeType: metadata.mimeType,
@@ -168,29 +274,6 @@ export class Card extends Component<CardProps, CardState> {
                 size: metadata.size,
               });
             }
-          }
-
-          switch (fileState.status) {
-            case 'uploading':
-              progress = fileState.progress;
-              status = 'uploading';
-              break;
-            case 'processing':
-              if (dataURI) {
-                status = 'complete';
-                progress = 1;
-              } else {
-                status = 'processing';
-              }
-              break;
-            case 'processed':
-              status = 'complete';
-              break;
-            case 'failed-processing':
-              status = 'failed-processing';
-              break;
-            case 'error':
-              status = 'error';
           }
 
           const shouldFetchRemotePreview =
@@ -210,7 +293,7 @@ export class Card extends Component<CardProps, CardState> {
             try {
               const mode =
                 resizeMode === 'stretchy-fit' ? 'full-fit' : resizeMode;
-              const blob = await mediaClient.getImage(resolvedId, {
+              const blob = await mediaClient.getImage(this.resolvedId, {
                 collection: collectionName,
                 mode,
                 height,
@@ -220,7 +303,7 @@ export class Card extends Component<CardProps, CardState> {
               dataURI = URL.createObjectURL(blob);
               if (contextId) {
                 dataURI = addFileAttrsToUrl(dataURI, {
-                  id: resolvedId,
+                  id: this.resolvedId,
                   collection: collectionName,
                   contextId,
                   mimeType: metadata.mimeType,
@@ -236,6 +319,17 @@ export class Card extends Component<CardProps, CardState> {
             }
           }
 
+          status = getCardStatusFromFileState(fileState, dataURI);
+          progress =
+            getCardProgressFromFileState(fileState, dataURI) || progress;
+
+          this.fireLoadingStatusAnalyticsEvent({
+            resolvedId: this.resolvedId,
+            status,
+            fileState,
+            metadata,
+          });
+
           this.notifyStateChange({
             metadata,
             status,
@@ -245,10 +339,69 @@ export class Card extends Component<CardProps, CardState> {
           });
         },
         error: error => {
+          this.fireLoadingStatusAnalyticsEvent({
+            resolvedId: this.resolvedId,
+            status: 'error',
+            error,
+          });
           this.notifyStateChange({ error, status: 'error' });
         },
       });
   }
+
+  shouldFireAnalyticsEvent = (
+    action: AnalyticsLoadingAction,
+    errorState: AnalyticsErrorStateAttributes,
+  ) => {
+    const previousFailReason =
+      this.lastErrorState && this.lastErrorState.failReason;
+    const previousErrorMessage =
+      this.lastErrorState && this.lastErrorState.error;
+
+    const isDifferentErrorState =
+      errorState.failReason !== previousFailReason ||
+      errorState.error !== previousErrorMessage;
+
+    const isDifferentAction = action !== this.lastAction;
+
+    return isDifferentAction || isDifferentErrorState;
+  };
+
+  fireLoadingStatusAnalyticsEvent = ({
+    resolvedId,
+    status,
+    fileState,
+    metadata,
+    error,
+  }: {
+    resolvedId: string;
+    status: CardStatus;
+    fileState?: FileState;
+    metadata?: FileDetails;
+    error?: Error;
+  }) => {
+    const { createAnalyticsEvent } = this.props;
+    const action = getAnalyticsStatusFromCardStatus(status);
+    const errorState = getAnalyticsErrorStateAttributes(fileState, error);
+
+    if (action && this.shouldFireAnalyticsEvent(action, errorState)) {
+      this.lastAction = action;
+      this.lastErrorState = errorState;
+      createAndFireCustomMediaEvent(
+        {
+          eventType: 'operational',
+          action,
+          actionSubject: 'mediaCardRender',
+          actionSubjectId: resolvedId,
+          attributes: {
+            fileAttributes: getFileAttributes(metadata),
+            ...errorState,
+          },
+        },
+        createAnalyticsEvent,
+      );
+    }
+  };
 
   notifyStateChange = (state: Partial<CardState>) => {
     if (this.hasBeenMounted) {
@@ -267,12 +420,15 @@ export class Card extends Component<CardProps, CardState> {
     if (this.hasBeenMounted) {
       this.setState({ dataURI: undefined });
     }
+    this.lastAction = undefined;
+    this.lastErrorState = {};
   };
 
   // This method is called when card fails and user press 'Retry'
   private onRetry = () => {
     const { identifier, mediaClient } = this.props;
-
+    this.lastAction = undefined;
+    this.lastErrorState = {};
     this.subscribe(identifier, mediaClient);
   };
 
@@ -355,6 +511,7 @@ export class Card extends Component<CardProps, CardState> {
         onError={this.onInlinePlayerError}
         onClick={this.onClick}
         selected={selected}
+        ref={this.cardRef}
       />
     );
   };
@@ -365,7 +522,7 @@ export class Card extends Component<CardProps, CardState> {
     });
   };
 
-  renderMediaViewer = () => {
+  renderMediaViewer = (): React.ReactPortal | undefined => {
     const { mediaViewerSelectedItem } = this.state;
     const { mediaClient, identifier, mediaViewerDataSource } = this.props;
     if (!mediaViewerSelectedItem) {
@@ -405,6 +562,7 @@ export class Card extends Component<CardProps, CardState> {
     const { progress, metadata, dataURI, previewOrientation } = this.state;
     const { onRetry, onCardViewClick, actions, onMouseEnter } = this;
     const status = getCardStatus(this.state, this.props);
+
     const card = (
       <CardView
         status={status}
@@ -423,6 +581,7 @@ export class Card extends Component<CardProps, CardState> {
         progress={progress}
         onRetry={onRetry}
         previewOrientation={previewOrientation}
+        ref={this.cardRef}
       />
     );
 
@@ -456,18 +615,12 @@ export class Card extends Component<CardProps, CardState> {
   render() {
     const { metadata } = this.state;
     return (
-      /* 
-        First Context provides data needed to build packageHierarchy in Atlaskit Analytics Listener and Media Analytics Listener.
-        This data is not added to the final GASv3 payload 
-      */
-      <AnalyticsContext data={getBaseAnalyticsContext()}>
-        {/* 
+      /*
           Second context provides data to be merged with any other context down in the tree and the event's payload.
           This data is usually not available at the time of firing the event, though it is needed to be sent to the backend.
-       */}
-        <AnalyticsContext data={getUIAnalyticsContext(metadata)}>
-          {this.renderContent()}
-        </AnalyticsContext>
+       */
+      <AnalyticsContext data={getUIAnalyticsContext(this.resolvedId, metadata)}>
+        {this.renderContent()}
       </AnalyticsContext>
     );
   }
@@ -506,3 +659,13 @@ export class Card extends Component<CardProps, CardState> {
     }
   };
 }
+
+/*
+  This Context provides data needed to build packageHierarchy in Atlaskit Analytics Listener and Media Analytics Listener.
+  This data is not added to the final GASv3 payload
+*/
+export const Card: React.ComponentType<
+  CardWithAnalyticsEventsProps
+> = withAnalyticsContext(getBaseAnalyticsContext())(
+  withAnalyticsEvents()(CardBase),
+);