+++ packages/media/media-picker/src/service/uploadServiceImpl.ts
@@ -7,6 +7,7 @@ import {
   MediaClient,
   globalMediaEventEmitter,
 } from '@atlaskit/media-client';
+import { RECENTS_COLLECTION } from '@atlaskit/media-client/constants';
 import {
   MediaStore,
   MediaStoreCopyFileWithTokenBody,
@@ -20,7 +21,6 @@ import {
 import { EventEmitter2 } from 'eventemitter2';
 import { MediaFile } from '../domain/file';
 
-import { RECENTS_COLLECTION } from '../popup/config';
 import { mapAuthToSourceFileOwner } from '../popup/domain/source-file';
 import { getPreviewFromImage } from '../util/getPreviewFromImage';
 import { UploadParams } from '..';